Add files via upload
Example of the system of Beta-2-adrenoreceptor (B2ar) prepared by Charmm-gui web interphace for GROMACS.
The model of B2ar was prepared based on its X-ray structure (pdb id 2rh1) omitting fragments of unresolved intracellular loop 3 (IC3). 
Molecular composition of the system: 1 GPCR, 120 lipids, 11729 water molecules and 155 ions (corresponded to physiological concentration of NaCl) This system can be used as tutorial of the modeling of membrane receptors with Gromacs. Inside the folder README file can be used as shell script to run equilibration and production run of the system in step-by-step routine.